# openapi.yaml
openapi: 3.1.0
info:
  title: CDDO Data Marketplace API
  version: 0.1.0
paths:
  /organisations:
    get:
      summary: List Organisations
      operationId: list_organisations_organisations_get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Organisation'
                type: array
                title: Response List Organisations Organisations Get
  /catalogue:
    get:
      summary: Search Catalogue
      operationId: search_catalogue_catalogue_get
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
            default: ""
            title: Query
        - name: topic
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
            default: []
            title: Topic
        - name: organisation
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/organisationID'
            default: []
            title: Organisation
        - name: assetType
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/assetType'
            default: []
            title: Assettype
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            title: Limit
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Offset
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAssetsResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /catalogue/{asset_id}:
    get:
      summary: Catalogue Entry Detail
      operationId: catalogue_entry_detail_catalogue__asset_id__get
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Asset Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDetailResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /publish/batch/create_job:
    post:
      summary: Create Batch Publish Job
      operationId: create_batch_publish_job_publish_batch_create_job_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_batch_publish_job_publish_batch_create_job_post'
        required: true
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeginBatchPublishJobResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /publish/batch/{jobID}/status:
    get:
      summary: Get Publish Job Status
      operationId: get_publish_job_status_publish_batch__jobID__status_get
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Jobid
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPublishJob'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /publish/batch/{jobID}/publish:
    post:
      summary: Publish Draft
      operationId: publish_draft_publish_batch__jobID__publish_post
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Jobid
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPublishJob'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /publish/batch/{jobID}/abort:
    post:
      summary: Abort Job
      operationId: abort_job_publish_batch__jobID__abort_post
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Jobid
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPublishJob'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AssetDetailResponse:
      properties:
        asset:
          anyOf:
            - $ref: '#/components/schemas/DatasetResponse'
            - $ref: '#/components/schemas/DataServiceResponse'
          title: Asset
      type: object
      required:
        - asset
      title: AssetDetailResponse
    BatchPublishJob:
      properties:
        jobID:
          type: string
          format: uuid
          title: Jobid
        jobStatus:
          $ref: '#/components/schemas/publishJobStatus'
      type: object
      required:
        - jobID
        - jobStatus
      title: BatchPublishJob
    BeginBatchPublishJobResponse:
      properties:
        jobID:
          type: string
          format: uuid
          title: Jobid
        jobStatus:
          $ref: '#/components/schemas/publishJobStatus'
        data:
          items:
            anyOf:
              - $ref: '#/components/schemas/CreateDatasetBody'
              - $ref: '#/components/schemas/CreateDataServiceBody'
          type: array
          title: Data
      type: object
      required:
        - jobID
        - jobStatus
        - data
      title: BeginBatchPublishJobResponse
    Body_create_batch_publish_job_publish_batch_create_job_post:
      properties:
        datasets:
          type: string
          format: binary
          title: Datasets
          description: The Dataset tab from the spreadsheet template in CSV format
        dataservices:
          type: string
          format: binary
          title: Dataservices
          description: The DataService tab from the spreadsheet template in CSV format
      type: object
      required:
        - datasets
        - dataservices
      title: Body_create_batch_publish_job_publish_batch_create_job_post
    ContactPoint:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        telephone:
          anyOf:
            - type: string
            - type: "null"
          title: Telephone
        address:
          anyOf:
            - type: string
            - type: "null"
          title: Address
      type: object
      required:
        - name
        - email
      title: ContactPoint
    CreateDataServiceBody:
      properties:
        created:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created
        summary:
          anyOf:
            - type: string
            - type: "null"
          title: Summary
        modified:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Modified
        title:
          type: string
          title: Title
        type:
          $ref: '#/components/schemas/assetType'
        accessRights:
          anyOf:
            - $ref: '#/components/schemas/rightsStatement'
            - type: "null"
        alternativeTitle:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Alternativetitle
          default: []
        contactPoint:
          $ref: '#/components/schemas/ContactPoint'
        issued:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Issued
        keyword:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Keyword
          default: []
        licence:
          anyOf:
            - type: string
              minLength: 1
              format: uri
            - type: "null"
          title: Licence
          default: https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
        relatedAssets:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Relatedassets
          default: []
        securityClassification:
          anyOf:
            - $ref: '#/components/schemas/securityClass'
            - type: "null"
          default: OFFICIAL
        theme:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Theme
          default: []
        version:
          anyOf:
            - type: string
            - type: "null"
          title: Version
          default: "1.0"
        endpointDescription:
          type: string
          title: Endpointdescription
        endpointURL:
          anyOf:
            - type: string
            - type: "null"
          title: Endpointurl
        servesData:
          items:
            type: string
            minLength: 1
            format: uri
          type: array
          title: Servesdata
        serviceStatus:
          $ref: '#/components/schemas/ServiceStatus'
        serviceType:
          $ref: '#/components/schemas/ServiceType'
        organisationID:
          $ref: '#/components/schemas/organisationID'
        creatorID:
          $ref: '#/components/schemas/organisationID'
      type: object
      required:
        - title
        - type
        - contactPoint
        - endpointDescription
        - servesData
        - serviceStatus
        - serviceType
        - organisationID
        - creatorID
      title: CreateDataServiceBody
    CreateDatasetBody:
      properties:
        created:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created
        summary:
          anyOf:
            - type: string
            - type: "null"
          title: Summary
        modified:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Modified
        title:
          type: string
          title: Title
        type:
          $ref: '#/components/schemas/assetType'
        accessRights:
          anyOf:
            - $ref: '#/components/schemas/rightsStatement'
            - type: "null"
        alternativeTitle:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Alternativetitle
          default: []
        contactPoint:
          $ref: '#/components/schemas/ContactPoint'
        issued:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Issued
        keyword:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Keyword
          default: []
        licence:
          anyOf:
            - type: string
              minLength: 1
              format: uri
            - type: "null"
          title: Licence
          default: https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
        relatedAssets:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Relatedassets
          default: []
        securityClassification:
          anyOf:
            - $ref: '#/components/schemas/securityClass'
            - type: "null"
          default: OFFICIAL
        theme:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Theme
          default: []
        version:
          anyOf:
            - type: string
            - type: "null"
          title: Version
          default: "1.0"
        distributions:
          items:
            $ref: '#/components/schemas/DistributionSummary'
          type: array
          title: Distributions
        updateFrequency:
          type: string
          title: Updatefrequency
        organisationID:
          $ref: '#/components/schemas/organisationID'
        creatorID:
          $ref: '#/components/schemas/organisationID'
      type: object
      required:
        - title
        - type
        - contactPoint
        - distributions
        - updateFrequency
        - organisationID
        - creatorID
      title: CreateDatasetBody
    DataServiceResponse:
      properties:
        catalogueCreated:
          type: string
          format: date-time
          title: Cataloguecreated
        catalogueModified:
          type: string
          format: date-time
          title: Cataloguemodified
        creator:
          $ref: '#/components/schemas/Organisation'
        identifier:
          type: string
          format: uuid
          title: Identifier
        organisation:
          $ref: '#/components/schemas/Organisation'
        created:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created
        summary:
          anyOf:
            - type: string
            - type: "null"
          title: Summary
        modified:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Modified
        title:
          type: string
          title: Title
        type:
          const: DataService
          title: Type
        accessRights:
          anyOf:
            - $ref: '#/components/schemas/rightsStatement'
            - type: "null"
        alternativeTitle:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Alternativetitle
          default: []
        contactPoint:
          $ref: '#/components/schemas/ContactPoint'
        issued:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Issued
        keyword:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Keyword
          default: []
        licence:
          anyOf:
            - type: string
              minLength: 1
              format: uri
            - type: "null"
          title: Licence
          default: https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
        relatedAssets:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Relatedassets
          default: []
        securityClassification:
          anyOf:
            - $ref: '#/components/schemas/securityClass'
            - type: "null"
          default: OFFICIAL
        theme:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Theme
          default: []
        version:
          anyOf:
            - type: string
            - type: "null"
          title: Version
          default: "1.0"
        endpointDescription:
          type: string
          title: Endpointdescription
        endpointURL:
          anyOf:
            - type: string
            - type: "null"
          title: Endpointurl
        servesData:
          items:
            type: string
            minLength: 1
            format: uri
          type: array
          title: Servesdata
        serviceStatus:
          $ref: '#/components/schemas/ServiceStatus'
        serviceType:
          $ref: '#/components/schemas/ServiceType'
      type: object
      required:
        - catalogueCreated
        - catalogueModified
        - creator
        - identifier
        - organisation
        - title
        - type
        - contactPoint
        - endpointDescription
        - servesData
        - serviceStatus
        - serviceType
      title: DataServiceResponse
      examples:
        - accessRights: INTERNAL
          catalogueCreated: "2023-07-31T11:26:50.383Z"
          catalogueModified: "2023-07-31T11:26:50.383Z"
          contactPoint:
            contactName: DWP Integration Team
            email: integration.technologyplatforms@dwp.gsi.gov.uk
          creator:
            acronym: DWP
            homepage: https://www.gov.uk/government/organisations/department-for-work-pensions
            id: department-for-work-pensions
            title: Department for Work & Pensions
          description: The location-service provides endpoints to perform a range of address based queries for UK locations....
          endpointDescription: https://engineering.dwp.gov.uk/apis/docs
          endpointURL: ""
          identifier: fcbc4d3f-0c05-4857-b0e3-eeec6bfea3a1
          issued: "2022-01-23"
          keyword:
            - Address Search
            - UPRN
          licence: https://opensource.org/license/isc-license-txt/
          modified: "2023-01-30"
          organisation:
            acronym: DWP
            homepage: https://www.gov.uk/government/organisations/department-for-work-pensions
            id: department-for-work-pensions
            title: Department for Work & Pensions
          relatedAssets: []
          securityClassification: OFFICIAL
          servesData:
            - https://www.data.gov.uk/dataset/2dfb82b4-741a-4b93-807e-11abb4bb0875/os-postcodes-data
            - https://www.data.gov.uk/dataset/03d48dba-529b-4bd5-93a5-6d41d1b20ff9/national-address-gazetteer
            - https://www.data.gov.uk/dataset/92b32629-8ad4-43cb-9952-7d104971fa12/one-scotland-gazetteer
          serviceStatus: LIVE
          serviceType: REST
          summary: DWP single strategic solution for looking up addresses including fuzzy search and UPRN.
          theme:
            - https://www.data.gov.uk/search?filters%5Btopic%5D=Mapping
          title: Address Lookup
          type: DataService
          version: 2.0.0
    DataServiceSummary:
      properties:
        catalogueCreated:
          type: string
          format: date-time
          title: Cataloguecreated
        catalogueModified:
          type: string
          format: date-time
          title: Cataloguemodified
        creator:
          $ref: '#/components/schemas/Organisation'
        identifier:
          type: string
          format: uuid
          title: Identifier
        organisation:
          $ref: '#/components/schemas/Organisation'
        created:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created
        summary:
          type: string
          title: Summary
        modified:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Modified
        title:
          type: string
          title: Title
        type:
          const: DataService
          title: Type
        serviceType:
          $ref: '#/components/schemas/ServiceType'
      type: object
      required:
        - catalogueCreated
        - catalogueModified
        - creator
        - identifier
        - organisation
        - summary
        - title
        - type
        - serviceType
      title: DataServiceSummary
    DatasetResponse:
      properties:
        catalogueCreated:
          type: string
          format: date-time
          title: Cataloguecreated
        catalogueModified:
          type: string
          format: date-time
          title: Cataloguemodified
        creator:
          $ref: '#/components/schemas/Organisation'
        identifier:
          type: string
          format: uuid
          title: Identifier
        organisation:
          $ref: '#/components/schemas/Organisation'
        created:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created
        summary:
          anyOf:
            - type: string
            - type: "null"
          title: Summary
        modified:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Modified
        title:
          type: string
          title: Title
        type:
          const: Dataset
          title: Type
        accessRights:
          anyOf:
            - $ref: '#/components/schemas/rightsStatement'
            - type: "null"
        alternativeTitle:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Alternativetitle
          default: []
        contactPoint:
          $ref: '#/components/schemas/ContactPoint'
        issued:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Issued
        keyword:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Keyword
          default: []
        licence:
          anyOf:
            - type: string
              minLength: 1
              format: uri
            - type: "null"
          title: Licence
          default: https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
        relatedAssets:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Relatedassets
          default: []
        securityClassification:
          anyOf:
            - $ref: '#/components/schemas/securityClass'
            - type: "null"
          default: OFFICIAL
        theme:
          anyOf:
            - items:
                type: string
                minLength: 1
                format: uri
              type: array
            - type: "null"
          title: Theme
          default: []
        version:
          anyOf:
            - type: string
            - type: "null"
          title: Version
          default: "1.0"
        distributions:
          items:
            $ref: '#/components/schemas/DistributionSummary'
          type: array
          title: Distributions
        updateFrequency:
          type: string
          title: Updatefrequency
      type: object
      required:
        - catalogueCreated
        - catalogueModified
        - creator
        - identifier
        - organisation
        - title
        - type
        - contactPoint
        - distributions
        - updateFrequency
      title: DatasetResponse
      examples:
        - accessRights: INTERNAL
          catalogueCreated: "2023-07-31T11:26:50.383Z"
          catalogueModified: "2023-07-31T11:26:50.383Z"
          contactPoint:
            contactName: DWP Integration Team
            email: integration.technologyplatforms@dwp.gsi.gov.uk
          creator:
            acronym: DWP
            homepage: https://www.gov.uk/government/organisations/department-for-work-pensions
            id: department-for-work-pensions
            title: Department for Work & Pensions
          description: The location-service provides endpoints to perform a range of address based queries for UK locations....
          distributions:
            - mediaType: CSV
              modified: "2023-08-01"
              title: document1
            - mediaType: XLS
              modified: "2023-08-01"
              title: document2
          identifier: fcbc4d3f-0c05-4857-b0e3-eeec6bfea3a1
          issued: "2022-01-23"
          keyword:
            - Address Search
            - UPRN
          licence: https://opensource.org/license/isc-license-txt/
          modified: "2023-01-30"
          organisation:
            acronym: DWP
            homepage: https://www.gov.uk/government/organisations/department-for-work-pensions
            id: department-for-work-pensions
            title: Department for Work & Pensions
          relatedAssets: []
          securityClassification: OFFICIAL
          summary: DWP single strategic solution for looking up addresses including fuzzy search and UPRN.
          theme:
            - https://www.data.gov.uk/search?filters%5Btopic%5D=Mapping
          title: Address Lookup
          type: Dataset
          updateFrequency: Monthly
          version: 2.0.0
    DatasetSummary:
      properties:
        catalogueCreated:
          type: string
          format: date-time
          title: Cataloguecreated
        catalogueModified:
          type: string
          format: date-time
          title: Cataloguemodified
        creator:
          $ref: '#/components/schemas/Organisation'
        identifier:
          type: string
          format: uuid
          title: Identifier
        organisation:
          $ref: '#/components/schemas/Organisation'
        created:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Created
        summary:
          type: string
          title: Summary
        modified:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Modified
        title:
          type: string
          title: Title
        type:
          const: Dataset
          title: Type
        mediaType:
          items:
            type: string
          type: array
          title: Mediatype
      type: object
      required:
        - catalogueCreated
        - catalogueModified
        - creator
        - identifier
        - organisation
        - summary
        - title
        - type
        - mediaType
      title: DatasetSummary
    DistributionSummary:
      properties:
        title:
          type: string
          title: Title
        modified:
          type: string
          format: date-time
          title: Modified
        mediaType:
          type: string
          title: Mediatype
        accessService:
          anyOf:
            - type: string
            - type: "null"
          title: Accessservice
        identifier:
          type: string
          title: Identifier
        issued:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Issued
        licence:
          anyOf:
            - type: string
              minLength: 1
              format: uri
            - type: "null"
          title: Licence
          default: https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
        byteSize:
          anyOf:
            - type: integer
            - type: "null"
          title: Bytesize
      type: object
      required:
        - title
        - modified
        - mediaType
        - identifier
      title: DistributionSummary
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Organisation:
      properties:
        title:
          type: string
          title: Title
        id:
          type: string
          title: Id
        acronym:
          type: string
          title: Acronym
        homepage:
          type: string
          minLength: 1
          format: uri
          title: Homepage
      type: object
      required:
        - title
        - id
        - acronym
        - homepage
      title: Organisation
    SearchAssetsResponse:
      properties:
        data:
          items:
            anyOf:
              - $ref: '#/components/schemas/DatasetSummary'
              - $ref: '#/components/schemas/DataServiceSummary'
          type: array
          title: Data
        facets:
          $ref: '#/components/schemas/SearchFacets'
      type: object
      required:
        - data
        - facets
      title: SearchAssetsResponse
    SearchFacet:
      properties:
        title:
          type: string
          title: Title
        id:
          type: string
          title: Id
      type: object
      required:
        - title
        - id
      title: SearchFacet
    SearchFacets:
      properties:
        topics:
          items:
            $ref: '#/components/schemas/SearchFacet'
          type: array
          title: Topics
        organisations:
          items:
            $ref: '#/components/schemas/SearchFacet'
          type: array
          title: Organisations
        assetTypes:
          items:
            $ref: '#/components/schemas/SearchFacet'
          type: array
          title: Assettypes
      type: object
      required:
        - topics
        - organisations
        - assetTypes
      title: SearchFacets
    ServiceStatus:
      type: string
      enum:
        - DISCOVERY
        - ALPHA
        - BETA
        - PRIVATE_BETA
        - PUBLIC_BETA
        - LIVE
        - DEPRECATED
        - WITHDRAWN
      title: ServiceStatus
    ServiceType:
      type: string
      enum:
        - REST
        - EVENT
        - SOAP
      title: ServiceType
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    assetType:
      type: string
      enum:
        - DataService
        - Dataset
      title: assetType
    organisationID:
      type: string
      enum:
        - department-for-work-pensions
        - food-standards-agency
        - nhs-digital
        - ordnance-survey
      title: organisationID
    publishJobStatus:
      type: string
      enum:
        - CREATED
        - DRAFT
        - PUBLISHED
        - ERROR
        - ABORTED
      title: publishJobStatus
    rightsStatement:
      type: string
      enum:
        - INTERNAL
        - OPEN
        - COMMERCIAL
      title: rightsStatement
    securityClass:
      type: string
      enum:
        - OFFICIAL
        - SECRET
        - TOP_SECRET
        - NOT_APPLICABLE
      title: securityClass